requirements.txt
pytest==6.2.5
argparse==1.4.0
pyyaml==6.0
pycryptodome==3.14.1
aiohttp==3.8.1
psycopg2-binary==2.8.6
aiohttp_jinja2==1.5
jinja2==3.0.3
pathlib2==2.3.7.post1
